requirements.txt
alembic==1.13.1
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Gladiator==0.7.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysqlclient==2.2.3
pillow==10.2.0
PyMySQL==1.1.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
Werkzeug==3.0.1
